search: add an indexer function to index a single file
This function will eventually be called by a larger, automated process
that will be introduced in future work.

Indexing a file involves:

* Passing in a docID that can be used if the blockpointer doesn't have
  one yet.
* Looking up the current blockpointer using the node metadata.
* Figuring out the content type of the document using mime extensions
  and/or content sniffing.
* Creating an indexable document struct for the document based on
  its content type, and populating that.
* Indexing that using Bleve.
* Recording the blockPtr -> docID and docID -> parentDocID mappings in
  local databases.

This also introduces some more general fixes and cleanups to get the
`Indexer` to work and clean up after itself in a test setting.

Issue: HOTPOT-1490

+++ go/kbfs/search/indexed_block_db.go
@@ -211,7 +211,7 @@ func (db *IndexedBlockDb) GetNextDocIDs(n int) ([]string, error) {
 // blockMD is per-block metadata for an individual indexed block.
 type blockMD struct {
 	// Exported only for serialization.
-	IndexVersion uint   `codec:"i"`
+	IndexVersion uint64 `codec:"i"`
 	DocID        string `codec:"d"`
 }
 
@@ -290,7 +290,7 @@ func (db *IndexedBlockDb) checkDbLocked(
 // Get returns the version and doc ID for the given block.
 func (db *IndexedBlockDb) Get(
 	ctx context.Context, ptr data.BlockPointer) (
-	indexVersion uint, docID string, err error) {
+	indexVersion uint64, docID string, err error) {
 	db.lock.RLock()
 	defer db.lock.RUnlock()
 	err = db.checkDbLocked(ctx, "IBD(Get)")
@@ -307,8 +307,8 @@ func (db *IndexedBlockDb) Get(
 
 // Put saves the version and doc ID for the given block.
 func (db *IndexedBlockDb) Put(
-	ctx context.Context, tlfID tlf.ID, ptr data.BlockPointer, indexVersion uint,
-	docID string) error {
+	ctx context.Context, tlfID tlf.ID, ptr data.BlockPointer,
+	indexVersion uint64, docID string) error {
 	db.lock.Lock()
 	defer db.lock.Unlock()
 	err := db.checkDbLocked(ctx, "IBD(Put)")
